+++ tests/test_stack_overflow.rs
@@ -6,7 +6,8 @@ use std::alloc::{alloc, Layout};
 
 use const_lru::ConstLru;
 
-type BigConstLru = ConstLru<usize, usize, 1_000_000>;
+// ~400 MB
+type BigConstLru = ConstLru<usize, usize, 10_000_000>;
 const BIG_CONST_LRU_LAYOUT: Layout = Layout::new::<BigConstLru>();
 
 /*
@@ -40,4 +41,16 @@ fn clear_doesnt_stack_overflow() {
     let mut c = boxed_big_const_lru();
     c.clear();
     assert!(c.insert(1, 2).is_none());
+}
+
+#[test]
+#[cfg_attr(miri, ignore)]
+fn clone_to_alloc_doesnt_stack_overflow() {
+    let c = boxed_big_const_lru();
+    let mut cloned = unsafe {
+        let new_alloc_ptr = alloc(BIG_CONST_LRU_LAYOUT) as *mut BigConstLru;
+        c.clone_to_alloc(new_alloc_ptr);
+        Box::from_raw(new_alloc_ptr)
+    };
+    assert!(cloned.insert(1, 2).is_none());
 }